+++ CHANGELOG.md
@@ -10,7 +10,8 @@
  * Gradle 5 compatible (up to 5.6.4) (#84)
  * Use Gradle 5.6.4 to build (#84)
  * Use Kotlin DSL 5.6.4 (#84)
- * Android Gradle Plugin 3.5.3 compatible (#86)
+ * Android Gradle Plugin 3.5.3 compatible (Kotlin 1.3.50) (#86)
+ * Android Gradle Plugin 3.6.4 compatible (Kotlin 1.3.72) (#98)
 
 ### Fixes
  * Make sure nothing is logged when lint task is disabled.